Undo moving as_xml to event as it is exporting occurences
