+++ src/onegov/event/collections/events.py
@@ -426,77 +426,6 @@ class EventCollection(Pagination):
         return self.from_import(items, publish_immediately=True,
                                 future_events_only=future_events_only)
 
-    def as_xml(self, future_events_only=True):
-        """
-        Returns all published occurrences as xml.
-
-        The xml format was Winterthur's wish (no specs behind). Their mobile
-        app will consume the events from xml
-
-        Format:
-        <events>
-            <event>
-                <id></id>
-                <title></title>
-                <tags></tags>
-                    <tag></tag>
-                <description></description>
-                <start></start>
-                <end></end>
-                <location></location>
-                <price></price>
-                ..
-            </event>
-            <event>
-                ..
-            </event>
-            ..
-        </events>
-
-        :param future_events_only: if set, only future events will be
-        returned, all events otherwise
-        :rtype: str
-        :return: xml string
-
-        """
-        xml = '<events></events>'
-        root = objectify.fromstring(xml)
-
-        query = self.session.query(Event)
-        for e in query:
-            # e = (self.session.query(Event)
-            #     .filter(Event.id == e.event_id).first())
-
-            if e.state != 'published':
-                continue
-            if future_events_only and datetime.fromisoformat(str(
-                    e.end)).date() < datetime.today().date():
-                continue
-
-            event = objectify.Element('event')
-            event.id = e.id
-            event.title = e.title
-            txs = tags(e.tags)
-            event.append(txs)
-            event.description = e.description
-            event.start = e.start
-            event.end = e.end
-            event.location = e.location
-            event.price = e.price
-            event.organizer = e.organizer
-            event.event_url = e.external_event_url
-            event.organizer_email = e.organizer_email
-            event.organizer_phone = e.organizer_phone
-            event.modified = e.last_change
-            root.append(event)
-
-        # remove lxml annotations
-        objectify.deannotate(root, pytype=True, xsi=True, xsi_nil=True)
-        etree.cleanup_namespaces(root)
-
-        return etree.tostring(root, encoding='utf-8', xml_declaration=True,
-                              pretty_print=True)
-
     def as_anthrazit_xml(self, request, future_events_only=True):
         """
         Returns all published occurrences as xml for Winterthur.
